flatten strategies support init
