--- src/flattenPath.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Keboola\Processor\FlattenFolders;
-
-/**
- * @param array $pathParts
- * @return string
- */
-function flattenPath(array $pathParts): string
-{
-    return join('-', array_map(function ($pathPart) {
-        return str_replace('-', '--', basename($pathPart));
-    }, $pathParts));
-}
